LayoutTests/fast/spatial-navigation/snav-imagemap-area-not-focusable-expected.txt




PASS gFocusedDocument.activeElement.getAttribute("id") is "1"
PASS gFocusedDocument.activeElement.getAttribute("id") is "3"
PASS gFocusedDocument.activeElement.getAttribute("id") is "1"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS successfullyParsed is true

TEST COMPLETE
This test tests that areas of an imagemap without an href are not focusable, thus can not be reached with spatial navigation.
